main.py
This is one of the convenient python projects that has a lot of use in the future. To create an Email slicer with python, users have to generate a program to retrieve the username and domain of the email.  You can even customize the application and send a message to the host with this information.